Add coreos-installer.target and make default
What we really want when in "installer mode" is to override the default
flow of the bootup. The cleanest way to do this is to define a new
default target for systemd rather than `initrd.target`. That way we
don't have to awkwardly wedge ourselves between existing services. One
practical consequence of this for example is that systemd no longer
prints "Waiting" messages, since we're freely running without
constraints.

Closes: #4