coreos-iso-embed-ignition: add tool to embed Ignition config in ISO
The only way to provide an Ignition config to live ISO boots is by
manually modifying the kernel command line, which is inconvenient.
To address this, https://github.com/coreos/coreos-assembler/pull/755
adds padding to the embedded initramfs image, and records in the
ISO 9660 System Area the offset and length of that padding within the
ISO.  Add a tool which takes a user-specified Ignition config, generates
a cpio.gz of it, and uses the recorded offset/length to embed it in a
live ISO image.  ignition-dracut will then ensure the config is read as
user.ign.